english-learning-flask/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask-WTF==1.2.1
email-validator==2.1.0.post1
youtube-transcript-api==0.6.1
requests==2.31.0
celery==5.3.6
redis==5.0.1
transformers>=4.35.2
torch>=2.2.0
Flask-HTMX==0.4.0
bcrypt==4.1.2
